requirements.txt
uvicorn==0.30.6
fastapi==0.112.1
pymongo==4.8.0
python-dotenv==1.0.1